src/main/webapp/app/entities/answer/list/answer.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpHeaders, HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { AnswerService } from '../service/answer.service';

import { AnswerComponent } from './answer.component';

describe('Answer Management Component', () => {
  let comp: AnswerComponent;
  let fixture: ComponentFixture<AnswerComponent>;
  let service: AnswerService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [RouterTestingModule.withRoutes([{ path: 'answer', component: AnswerComponent }]), HttpClientTestingModule, AnswerComponent],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            data: of({
              defaultSort: 'id,asc',
            }),
            queryParamMap: of(
              jest.requireActual('@angular/router').convertToParamMap({
                page: '1',
                size: '1',
                sort: 'id,desc',
              }),
            ),
            snapshot: { queryParams: {} },
          },
        },
      ],
    })
      .overrideTemplate(AnswerComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(AnswerComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(AnswerService);

    const headers = new HttpHeaders();
    jest.spyOn(service, 'query').mockReturnValue(
      of(
        new HttpResponse({
          body: [{ id: 123 }],
          headers,
        }),
      ),
    );
  });

  it('Should call load all on init', () => {
    // WHEN
    comp.ngOnInit();

    // THEN
    expect(service.query).toHaveBeenCalled();
    expect(comp.answers?.[0]).toEqual(expect.objectContaining({ id: 123 }));
  });

  describe('trackId', () => {
    it('Should forward to answerService', () => {
      const entity = { id: 123 };
      jest.spyOn(service, 'getAnswerIdentifier');
      const id = comp.trackId(0, entity);
      expect(service.getAnswerIdentifier).toHaveBeenCalledWith(entity);
      expect(id).toBe(entity.id);
    });
  });
});
